Add anchorJS for in-page links
This adds AnchorJS (https://www.bryanbraun.com/anchorjs/) for in-page links which makes it quick and easy to share links to specific sections of the docs with colleagues. It loads from a CDN and uses the default options and styling which work well. There's just a small change added to _styleguide.scss to disable the underline on hover as I didn't that was necessary. The behaviour/styling is very similar to Bootstrap's own docs (which uses the same library).
e.g. https://getbootstrap.com/docs/4.4/getting-started/introduction/