Adds icon after external links
For technical docs, blogs and articles it’s often a good idea to differentiate between links within the site (internal) and links to another site (external) as it can be jarring to be reading a blog or technical docs, clicking on a link and getting kicked to an external site unexpectedly.

A nice UX pattern is to display an ‘external link’ icon next to external links. A simple way to implement this is using a CSS pseudo selector targeting only links which begin with http. This PR addresses this.

You can preview the changes at: https://styles-dev.trimblemaps.com/